requirements.txt
opencv-python>=4.8.0
google-genai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
